Added World to comser, is used for grouping scenese and systems. Added partial multi threaded and multi window support for platform. Fixed unit tests, now running.

+++ Comser/EntityList.h
@@ -99,7 +99,7 @@ namespace Comser
              */
             EntityIterator              findComponent( EntityId id, LocalComponentType type );
 
-            unsigned int                size() const
+            size_t                size() const
             {
                 return _entities.size();
             }
